+++ apps/edit/CHANGELOG.md
@@ -24,6 +24,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.9.1] - 2026-01-05
+
+### Security
+
+- Minor upgrade for `react`.
+- Patch upgrade for `next`.
+
 ## [1.9.0] - 2025-01-16
 
 ### Changed